PR Fix: Failing authoring tests
