+++ templates/vue/src/components/node-modal/AppearanceForm.vue
@@ -4,7 +4,7 @@
     <b-form-group>
       <b-form-checkbox
         v-model="addThumbnail"
-        data-testid="node-appearance-add-thumbnail"
+        data-testid="node-appearance-thumbnail"
       >
         Add a thumbnail
       </b-form-checkbox>
@@ -12,14 +12,14 @@
     <b-form-group v-if="addThumbnail">
       <file-upload
         v-model="node.imageURL"
-        data-testid="node-imageUrl"
+        input-test-id="node-appearance-thumbnail-url"
         placeholder="Enter the URL for the thumbnail"
       />
     </b-form-group>
     <b-form-group v-if="addThumbnail">
       <b-form-checkbox
         v-model="addLockedThumbnail"
-        data-testid="node-appearance-add-locked-thumbnail"
+        data-testid="node-appearance-lockedThumbnail"
       >
         Show a different thumbnail when locked
       </b-form-checkbox>
@@ -27,14 +27,14 @@
     <b-form-group v-if="addThumbnail && addLockedThumbnail">
       <file-upload
         v-model="node.lockedImageURL"
-        data-testid="node-lockedImageURL"
+        input-test-id="node-lockedImageURL"
         placeholder="Enter the URL for the thumbnail"
       />
     </b-form-group>
     <b-form-group>
       <b-form-checkbox
         v-model="node.hideTitle"
-        data-testid="node-appearance-hide-title"
+        data-testid="node-appearance-hideTitle"
       >
         Hide node title
       </b-form-checkbox>
@@ -42,7 +42,7 @@
     <b-form-group>
       <b-form-checkbox
         v-model="node.hideProgress"
-        data-testid="node-appearance-hide-progress"
+        data-testid="node-appearance-hideProgress"
       >
         Hide progress bar
       </b-form-checkbox>
@@ -50,7 +50,7 @@
     <b-form-group>
       <b-form-checkbox
         v-model="node.hideMedia"
-        data-testid="node-appearance-hide-media"
+        data-testid="node-appearance-hideMedia"
       >
         Hide media button
       </b-form-checkbox>
