+++ templates/vue/src/components/node-modal/content-form/VideoForm.vue
@@ -4,7 +4,7 @@
       <file-upload
         id="node-video-media-url"
         v-model="node.typeData.mediaURL"
-        data-testid="node-videoUrl"
+        input-test-id="node-videoUrl"
         placeholder="Enter URL for MP4 or YouTube video"
         required
       />
